refactor(MovieItems): replaceSkeleton 메서드 MovieItem으로 역할 분리
